refactor: relocate pagination
